fix: Moved importer code to webpack.util to share between screen shot test and main build.
