+++ packages/webpack.config.js
@@ -18,31 +18,8 @@ const webpack = require('webpack');
 const {readdirSync, lstatSync} = require('fs');
 const path = require('path');
 const ExtractTextPlugin = require('extract-text-webpack-plugin');
-const sass = require('node-sass');
 
-function tryResolve_(url, sourceFilename) {
-  // Put require.resolve in a try/catch to avoid node-sass failing with cryptic libsass errors when the importer throws
-  try {
-    return require.resolve(url, {paths: [path.dirname(sourceFilename)]});
-  } catch (e) {
-    return '';
-  }
-}
-
-function tryResolveScss(url, sourceFilename) {
-  // Support omission of .scss and leading _
-  const normalizedUrl = url.endsWith('.scss') ? url : `${url}.scss`;
-  return tryResolve_(normalizedUrl, sourceFilename) ||
-    tryResolve_(path.join(path.dirname(normalizedUrl), `_${path.basename(normalizedUrl)}`), sourceFilename);
-}
-
-function importer(url, prev) {
-  if (url.startsWith('@material')) {
-    const resolved = tryResolveScss(url, prev);
-    return {file: resolved || url};
-  }
-  return {file: url};
-}
+const {importer} = require('./webpack.util');
 
 const isDirectory = (source) => lstatSync(source).isDirectory();
 const containsJsFile = (source) => readdirSync(source).some((file) => path.extname(file) === '.js');
@@ -137,9 +114,7 @@ function getCssWebpackConfig(entryPath, chunk) {
               },
             }, {
               loader: 'sass-loader',
-              options: {
-                importer: importer,
-              },
+              options: {importer},
             }],
           }),
         }],
