+++ test/screenshot/webpack-bundles.js
@@ -1,32 +1,7 @@
 const ExtractTextPlugin = require('extract-text-webpack-plugin');
 const OptimizeCssAssetsPlugin = require('optimize-css-assets-webpack-plugin');
 
-const sass = require('node-sass');
-const path = require('path');
-
-function tryResolve_(url, sourceFilename) {
-  // Put require.resolve in a try/catch to avoid node-sass failing with cryptic libsass errors when the importer throws
-  try {
-    return require.resolve(url, {paths: [path.dirname(sourceFilename)]});
-  } catch (e) {
-    return '';
-  }
-}
-
-function tryResolveScss(url, sourceFilename) {
-  // Support omission of .scss and leading _
-  const normalizedUrl = url.endsWith('.scss') ? url : `${url}.scss`;
-  return tryResolve_(normalizedUrl, sourceFilename) ||
-    tryResolve_(path.join(path.dirname(normalizedUrl), `_${path.basename(normalizedUrl)}`), sourceFilename);
-}
-
-function importer(url, prev) {
-  if (url.startsWith('@material')) {
-    const resolved = tryResolveScss(url, prev);
-    return {file: resolved || url};
-  }
-  return {file: url};
-}
+const {importer} = require('../../packages/webpack.util');
 
 module.exports.bundle = function(testPath, outputPath) {
   return {
@@ -48,9 +23,7 @@ module.exports.bundle = function(testPath, outputPath) {
             },
             {
               loader: 'sass-loader',
-              options: {
-                importer: importer,
-              },
+              options: {importer},
             },
           ],
         }),
